docker/flask_api/requirements.txt
async-timeout==4.0.2
attrs==21.4.0
aws==0.2.5
awscli==1.18.69
backports.zoneinfo==0.2.1
bcrypt==3.2.0
beautifulsoup4==4.9.3
blis==0.7.8
boto==2.49.0
boto3==1.16.1
botocore==1.19.1
bs4==0.0.1
catalogue==2.0.8
certifi==2019.11.28
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
click==8.0.4
clickclick==20.10.2
colorama==0.4.3
connexion==2.7.0
cryptography==36.0.1
cymem==2.0.6
dateparser==1.1.1
Deprecated==1.2.13
docutils==0.16
ecdsa==0.17.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.0/en_core_web_sm-3.4.0-py3-none-any.whl
fabric==2.6.0
flasgger==0.9.5
Flask==2.0.1
Flask-Cors==3.0.10
future==0.18.2
google==3.0.0
idna==2.8
importlib-metadata==4.11.3
importlib-resources==5.4.0
inflection==0.5.1
invoke==1.6.0
itsdangerous==2.1.1
Jinja2==3.0.3
jmespath==0.9.4
jsonschema==4.4.0
langcodes==3.3.0
lxml==4.6.3
MarkupSafe==2.1.0
mistune==2.0.2
murmurhash==1.0.8
mysql-connector-python==8.0.22
numpy==1.19.2
olefile==0.46
openapi-spec-validator==0.2.9
packaging==21.3
pandas==1.1.3
paramiko==2.7.2
pathlib2==2.3.7.post1
pathy==0.6.2
pdf2image==1.14.0
Pillow==7.0.0
preshed==3.0.7
prettytable==3.2.0
probableparsing==0.0.1
protobuf==3.19.4
pyasn1==0.4.2
pycparser==2.21
pydantic==1.9.2
Pygments==2.3.1
PyNaCl==1.5.0
pyodbc==4.0.32
pyparsing==3.0.7
pyrsistent==0.18.1
python-apt==2.0.0+ubuntu0.20.4.8
python-crfsuite==0.9.7
python-dateutil==2.7.3
python-distutils-extra==2.39
python-dotenv==0.20.0
python-fullcontact==3.1.0
python-jose==3.3.0
python-json-logger==2.0.2
pytz==2020.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==5.3.1
redis==3.5.3
regex==2022.3.2
requests==2.23.0
requests-unixsocket==0.2.0
roman==2.0.0
rsa==4.0
s3transfer==0.3.3
selenium==3.141.0
simplejson==3.16.0
six==1.14.0
smart-open==5.2.1
soupsieve==2.3.1
spacy==3.4.1
spacy-legacy==3.0.10
spacy-loggers==1.0.3
srsly==2.4.4
stripe==4.0.0
swagger-ui-bundle==0.0.8
textract-trp==0.1.3
thinc==8.1.0
tqdm==4.64.0
typer==0.4.2
typing_extensions==4.3.0
tzdata==2022.2
tzlocal==4.2
urllib3==1.25.7
usaddress==0.5.10
uWSGI==2.0.19.1
wasabi==0.10.1
wcwidth==0.2.5
Werkzeug==2.0.1
wikipedia==1.4.0
wrapt==1.14.0
zipp==3.7.0
shortuuid==1.0.9
postal==1.1.10
openpyxl==3.0.10
apminsight==1.0.3
Flask-Session==0.4.0
logtail-python==0.1.3
